refactor: merge useProductStoreData into useProductStores
- Migrated all data states, actions, and API methods from `src/composables/useProductStoreData.ts` to `src/composables/useProductStores.ts` to maintain a single source of truth for the Product Store domain.
- Deleted the redundant `useProductStoreData.ts` file.
- Updated the call site in `src/views/ProductStoreOnboarding.vue` to point to the newly merged composable module.
